`@remotion/studio`: Split selected outline components
